Fix home page testing and add home POST testing
We've add a CSRF token at the home page and it breaks the test
because each time the page is rendered, a new CSRF token is generated
so we are neve able to get the same results.

This commit removes CSRF tokens from both the observed and the expected
HTML in order to make a proper comparison. It also adds a new unit
test to check if we're able to properly handle POST requests.

Sources:
https://www.reddit.com/r/learnpython/comments/3vjxzn/django_unit_testing_csrf_token_in_html_assertion/cy7432u